requirements.txt
google-api-python-client==2.108.0
google-auth==2.25.0
azure-storage-blob==12.19.0
snowflake-connector-python==3.7.0
python-dotenv==1.0.0
